requirements.txt
streamlit==1.21.0
matplotlib==3.7.1
Pillow==9.4.0
